updated to std::sqrtf

+++ src/ray.h
@@ -2,6 +2,9 @@
 #define RAYTRACING_RAY_H
 
 
+#include <cmath>
+
+
 #include "vec3.h"
 #include "color.h"
 
@@ -36,7 +39,7 @@ float hit_sphere(const Vec3 &center, float radius, const Ray &r)
     if (discriminant < 0)
         return -1.0f;
     else
-        return (-b - std::sqrt(discriminant)) / (2.0f * a);
+        return (-b - std::sqrtf(discriminant)) / (2.0f * a);
 }
